moved landing components
