+++ .eslintrc.json
@@ -76,7 +76,7 @@
 
     // make sure there is no :any in the project
     "@typescript-eslint/no-explicit-any": "warn",
-    
+
     "@typescript-eslint/triple-slash-reference": 0,
     "@typescript-eslint/strict-boolean-expressions": 0,
     "@typescript-eslint/space-before-function-paren": 0,
