+++ src/app/(landing)/_components/FAQs/const.ts
