+++ src/app/(landing)/_components/FAQs/index.tsx
