+++ src/app/(landing)/_components/FAQs/style.css
