+++ src/app/(landing)/_components/FAQs/types.ts
