+++ src/app/(landing)/_components/Features/types.ts
