+++ src/app/(landing)/_components/Footer/index.tsx
