+++ src/app/(landing)/_components/Footer/types.ts
