+++ src/app/(landing)/_components/Header/types.ts
