+++ src/app/(landing)/_components/Hero/types.ts
