+++ src/app/(landing)/_components/Mission/index.tsx
