+++ src/app/(landing)/_components/Navbar/types.ts
