+++ src/app/(landing)/_components/SignUp/index.tsx
@@ -1,10 +1,10 @@
 import { addToWaitlist, getSpotsLeft } from '@/actions/waitlist.actions'
 import React, { useCallback, useEffect, useState } from 'react'
 
-import Button from '@/components/Button'
+import Button from '@/app/(landing)/_components/Button'
 import Image from 'next/image'
 import Link from 'next/link'
-import { Checkbox } from '../ui/checkbox'
+import { Checkbox } from '../../../../components/ui/checkbox'
 import './style.css'
 
 const { NODE_ENV } = process.env
