+++ src/app/(landing)/page.tsx
@@ -1,15 +1,15 @@
 'use client'
 
-import Hero from '@/components/Hero'
-import Team from '@/components/Team'
-import Faqs from '@/components/FAQs'
+import Hero from '@/app/(landing)/_components/Hero'
+import Team from '@/app/(landing)/_components/Team'
+import Faqs from '@/app/(landing)/_components/FAQs'
 import { scrollToRef } from '@/hooks'
-import Header from '@/components/Header'
-import SignUp from '@/components/SignUp'
-import Footer from '@/components/Footer'
-import Mission from '@/components/Mission'
-import Project from '@/components/Project'
-import Features from '@/components/Features'
+import Header from '@/app/(landing)/_components/Header'
+import SignUp from '@/app/(landing)/_components/SignUp'
+import Footer from '@/app/(landing)/_components/Footer'
+import Mission from '@/app/(landing)/_components/Mission'
+import Project from '@/app/(landing)/_components/Project'
+import Features from '@/app/(landing)/_components/Features'
 import React, { useRef, useEffect, useState } from 'react'
 import { BrowserRouter as Router } from 'react-router-dom'
